pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ask-cli"
version = "2.0.0"
description = "Terminal AI assistant"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.79.0",
    "rich>=13.7.1",
    "pyyaml>=6.0.1",
    "openai>=1.0.0",
    "google-genai>=1.0.0",
    "pyperclip>=1.9.0",
]

[project.scripts]
ask = "ask_cli.cli:main"

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-cov>=5.0.0", "ruff>=0.4.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/ask_cli"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/ask_cli --cov-report=term-missing"
